Split actions on multi files
